Move the judge wrapper into sarvam_intent_entity/ package
All Sarvam intent/entity code now lives under stt/sarvam_intent_entity/:
the standalone stt/intent_entity.py becomes sarvam_intent_entity/judge.py.
metrics.py calls it via the package; tests patch it there.
